refactor(command): move admin and fn in their own packages
