+++ internal/command/fn/fn_test.go
@@ -14,7 +14,7 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
-package command
+package fn
 
 import (
 	"bytes"
